Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/fetch/security/redirect-to-url-with-credentials.https-expected.txt

CONSOLE MESSAGE: Cross-origin redirection to https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?ACAOrigin=%20https://localhost:9443&PNGIMAGE&ACACredentials=true denied by Cross-Origin Resource Sharing policy: redirection URL https://foo:bar@127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?ACAOrigin=%20https://localhost:9443&PNGIMAGE&ACACredentials=true has credentials
CONSOLE MESSAGE: Fetch API cannot load https://localhost:9443/common/redirect.py?location=https://foo:bar@127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?ACAOrigin=%20https://localhost:9443%26PNGIMAGE%26ACACredentials=true due to access control checks.
CONSOLE MESSAGE: Cross-origin redirection to https://127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?ACAOrigin=%20https://localhost:9443&PNGIMAGE&ACACredentials=true denied by Cross-Origin Resource Sharing policy: redirection URL https://foo:bar@127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?ACAOrigin=%20https://localhost:9443&PNGIMAGE&ACACredentials=true has credentials
CONSOLE MESSAGE: Cannot load image https://localhost:9443/common/redirect.py?location=https://foo:bar@127.0.0.1:9443/service-workers/service-worker/resources/fetch-access-control.py?ACAOrigin=%20https://localhost:9443%26PNGIMAGE%26ACACredentials=true due to access control checks.

PASS No CORS fetch after a redirect with an URL containing credentials 
PASS CORS fetch after a redirect with a cross origin URL containing credentials 
PASS CORS fetch after a redirect with a same origin URL containing credentials 
PASS Image loading after a redirect with an URL containing credentials 
PASS CORS Image loading after a redirect with a cross origin URL containing credentials 
PASS CORS Image loading after a redirect with a same origin URL containing credentials 
PASS Frame loading after a redirect with an URL containing credentials 
